restructured root-painter to src-layout
restructuring the root-painter to a regular python src layout which is understood by the python
tool-chain and configured project with dependencies in pyproject.toml. The relative imports has been
fixed to `root_painter.<module_name>` which enables the UI to be started as a callable from the
main, this is the `root-painter` script defined in the pyproject.toml.

The build scripts has been moved to a top-level folder *build/* as it is separate from the
source code.